Code Game Ping-Pong
Ping Pong is an Olympic sport, a sports game with a ball, in which special rackets and a game table are used, delimited by a grid in half. The game can be played between two opponents or two pairs of opponents. The task of the players is to keep the ball in the game with the help of rackets — each player, after one bounce of the ball on his half of the table, must send the ball to the opponent's half of the table. A point is awarded to a player or a pair of players when the opponent cannot return the ball in accordance with the rules. Each game lasts up to 11 points, the match consists of an odd number of games, and is played for the most wins in the games.

1)The meaning of this game is not to let the ball touch the left and right sides of the screen.
2)If the ball touches the left screen, then 1 player will win if the right one, then the second one.
3)The left racket is controlled using the W and S keys, and the right one using PAGE_UP and PAGE_DOWN.